backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
tqdm
pypdf
faiss-cpu
numpy
python-dotenv
pydantic
requests
